Set node radius proportional to its quantity of children
Set the radius of each node proportional to the quantity of children it has.
Added separation between two node brothers when they are not collapsed (so they do not overlap due to the radii increase).
Added an icon for the root node and a function that draws it just in the root node.

BUG: image is not showing.