Center and zoom to fit initial visualization
Center the graph to the canvas and zoom it (in or out) to fit in the canvas area.
Create functions to compute dimensions (**computeDimensions**) of the graph and to translate and scale
the visualization to fit the canvas (**zoomToFitGraph**) for the initial visualization.

Modify the function (**zoom**) that defines the behaviour of the zoom and translation.