WIP centre graph
Add/substract the radius of the node to the compute of the box sussounding our graph.
Change initial coordinates of the graph.

Before our root node was being drawn in the centre of the canvas (around 1100 pixels), but we want the centre of the graph
to be displayed in the centre of the canvas.

NOTE: the if-else. The values not set (newX or newY) would be 0 and displace our graph.