Solve OR bug + highlight edges if both endpoints are highlighted
Solve OR bug: if no tag is selected (with the OR property), keep all nodes and edges highlighted.

Highlight nodes according to selection.
If both endpoints of an edge are highlighted, highlight the edge.

Created individual functions to highlight by categories (toxicity, stance,...).
Created funtions to manage the AND/OR highlighting calling to the proper functions.

It will be faster in the future to add the checkboxes per category and give the multiple AND/OR submenus behaviour.