WIP avoid edge show over unhighlighted node
Create image of a circle with background colour (transparent outside the circle).
Append the image to the node centered and scaled with size to cover the whole node.

WIP: Modify funtions to unhighlight to try to just change the opacity of all the associated classes of the node (nodeCircle, nodeText,
all features and all targets) except the backgroundCircle.

Previous version: added a function to give to an unhighlighted node a different colour with opacity 1.