Change node radius and separation to reduce empty space
Extract the code to compute the radius of a node in a function.
Change the set of valid radius of a node to reduce the differences between them.
0 children: radius = 10; 1: 13, 2: 16, x>2: 16 + 4*(x-2)

Change the *separationHeight* that defines the area reserved for a node to 10 pixels.
Change the *separate* function to just take into account the radius of two nodes and separate them according to that.

BUG: there might be a problem centering graphs wider than tall.