Displace graphs wider than tall to the middle of the canvas
For graphs wider than tall, the graph was being displayed on the top of our canvas, cutting the upper part of the graph.
If the graph is wider, we can displace it (vertically) to the middle of the canvas. Since we are scaling it and it is wider than tall,
we know it will fit if we position it at the canvasHeight/2.

Add (horizontally) the radius of the root node to the X position so it does not get cut.

Some WIP trying to hide the edge over the nodes for unhighlighted elements.