Refactor: change to const with different scope + delete center functions
Change **var** variables to **const** constants.
Change the scope of the constants (outside the d3.json and at the beginning of the file).
Change some function parameters that use constants.
Change some variables to let (in loop fors ).

Delete unnecessary/unused functions to center a node and a link on click.