WIP: zoom to fit and center first visualization of radial graph
**First approach** for the initial visualization of the graph.

Add the radius of the node when computing the dimensions of our graph (in *computeDimensions()*, so the nodes do not appear *cut* in half).
Implement first approach to *zoomToFit()*.
Modify *zoom()* to take into account the first zoom and (x,y) positions to **avoid a jump** when the user zooms or translates for the first time.

NOTE: right now we are computing the maximum radius in each quadrant without taking into account the angle.
Thus, it is not perfectly centered nor scaled, since we are not using the rectangular coordinates.