Hierarchical (#26)
* Set node radius proportional to its quantity of children

Set the radius of each node proportional to the quantity of children it has.
Added separation between two node brothers when they are not collapsed (so they do not overlap due to the radii increase).
Added an icon for the root node and a function that draws it just in the root node.

BUG: image is not showing.

* Hide features appearing in root node

Note: since the name of the root node is not "News Article" the icons were being drawn.
Now we check: if the parent of the node is undefined, then (it is the root node) and should not display the icons.

* Resize rings for the target visualization

The width and height of the rings images are based on the radius of the node.
Rightnow, just 4 pixels of radius smaller.

* Change edge length and node separation

ToDo: Find a good ratio and recompute the return value of the function **separation** to avoid overlapping of nodes.

* Resize ring images with a 10% of radius difference

* Displace targets outside the node (based on node radius)

* Resize features as Trivial cheese

Resize images drawing the features as trivial cheese to be a 10% smaller than its node.

* Displace feature dots to avoid overlapping between them

NOTE: the code commented would displace the nodes to start at the margin right of the node
(i.e., the dot for argumentation is touching the right border of its node).

I believe they just wanted the dots not the overlap each other (not to avoid overlapping the node,
 what would cause the dots not to be in columns).

* Undo displacement of targets outside the node

* Reposition and resize _all in one_ feature visualization

Position images at the center of the node and resize them to be as big as the node.

Refactor functions *positionImage* and *sizeImage* to support default parameter value and rewrite the percentage as the formula is supposed to be.

* Solve bug: show image for root node

* Center and zoom to fit initial visualization

Center the graph to the canvas and zoom it (in or out) to fit in the canvas area.
Create functions to compute dimensions (**computeDimensions**) of the graph and to translate and scale
the visualization to fit the canvas (**zoomToFitGraph**) for the initial visualization.

Modify the function (**zoom**) that defines the behaviour of the zoom and translation.

* WIP: Position target icons on the node scaled in a triangle

Position the icons in a triangle and scaled to be the size of the radius of its node.

ToDo: icons for target Group and Stereotype are not entirely inside of the node, translate and/or rescale.

* WIP centre graph

Add/substract the radius of the node to the compute of the box sussounding our graph.
Change initial coordinates of the graph.

Before our root node was being drawn in the centre of the canvas (around 1100 pixels), but we want the centre of the graph
to be displayed in the centre of the canvas.

NOTE: the if-else. The values not set (newX or newY) would be 0 and displace our graph.

* Reposition images for target icons on the node

* Solve OR bug + highlight edges if both endpoints are highlighted

Solve OR bug: if no tag is selected (with the OR property), keep all nodes and edges highlighted.

Highlight nodes according to selection.
If both endpoints of an edge are highlighted, highlight the edge.

Created individual functions to highlight by categories (toxicity, stance,...).
Created funtions to manage the AND/OR highlighting calling to the proper functions.

It will be faster in the future to add the checkboxes per category and give the multiple AND/OR submenus behaviour.

* WIP avoid edge show over unhighlighted node

Create image of a circle with background colour (transparent outside the circle).
Append the image to the node centered and scaled with size to cover the whole node.

WIP: Modify funtions to unhighlight to try to just change the opacity of all the associated classes of the node (nodeCircle, nodeText,
all features and all targets) except the backgroundCircle.

Previous version: added a function to give to an unhighlighted node a different colour with opacity 1.

* Change node radius and separation to reduce empty space

Extract the code to compute the radius of a node in a function.
Change the set of valid radius of a node to reduce the differences between them.
0 children: radius = 10; 1: 13, 2: 16, x>2: 16 + 4*(x-2)

Change the *separationHeight* that defines the area reserved for a node to 10 pixels.
Change the *separate* function to just take into account the radius of two nodes and separate them according to that.

BUG: there might be a problem centering graphs wider than tall.

* Change radius factor + reduce space between brother nodes.

Change the factor by which we multiply the number fo children to 3.
Reduce the separation between two node brothers to 5 pixels.
(See the difference for two brothers of radius 10 and the separation between them).

* Displace graphs wider than tall to the middle of the canvas

For graphs wider than tall, the graph was being displayed on the top of our canvas, cutting the upper part of the graph.
If the graph is wider, we can displace it (vertically) to the middle of the canvas. Since we are scaling it and it is wider than tall,
we know it will fit if we position it at the canvasHeight/2.

Add (horizontally) the radius of the root node to the X position so it does not get cut.

Some WIP trying to hide the edge over the nodes for unhighlighted elements.

* Displace targets and features drawn outside the node so they do not overlap with it