Attempt to improve the handling and creation of layouts and nodes.
- Fixed typo in radial_graph.js file.

In the Tree, Force and Radial layouts, clicking on a node or moving it blocked the screen for a few seconds. I discovered that this was due to the Update function being called every time a node was dragged or clicked. This also meant that you could not double click on a node. For now I have applied the solution only to the Force Layout, since it was where these problems were more noticeable as the screen was blocked more time and was much less fluid.

    - Now the update function is only called if a single click is made, it is not called if a node is dragged or double-clicked.
    - I have added a custom function for the double click event, so that the node is defixed.

Moving nodes and moving around the graph is not very smooth. This behavior is more noticeable with Circular Glyphs than with Dots. I believe that, in part, it is due to the fact that whoever designed the Features of Dots and Circular Glyphs implemented them in such a way that all the SVG images of each color of the possible options to be highlighted were loaded, and only changing the value of the opacity of these to be shown or not. This made that in all the graph all those images were loaded and moving with each node, although they could not be seen.

    - To solve it I have done that the images that should not be seen, simply are not loaded and therefore now it goes much more fluid.